% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_annotations}
\alias{example_annotations}
\title{Example Annotation Data}
\format{
A data frame with 20 rows and 5 columns:
\describe{
\item{Probe.Set.ID}{SNP identifier}
\item{Chromosome}{Chromosome location}
\item{Physical.Position}{Base pair position}
\item{Allele.A}{First allele (deletion marker "-" for indels)}
\item{Allele.B}{Second allele (insertion sequence for indels)}
}
}
\usage{
example_annotations
}
\description{
A dataset containing annotation information for 20 SNPs, including 10 regular SNPs and 10 indels.
}
\examples{
data(example_annotations)
head(example_annotations)
}
\keyword{datasets}
