% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_genotypes}
\alias{example_genotypes}
\title{Example Genotype Data}
\format{
A data frame with 20 rows and 4 columns:
\describe{
\item{probeset_id}{SNP identifier}
\item{Sample1}{Genotype codes (0, 1, 2) for Sample1}
\item{Sample2}{Genotype codes (0, 1, 2) for Sample2}
\item{Sample3}{Genotype codes (0, 1, 2) for Sample3}
}
}
\usage{
example_genotypes
}
\description{
A dataset containing example genotype data for 20 SNPs, including 10 indels.
}
\examples{
data(example_genotypes)
head(example_genotypes)
}
\keyword{datasets}
