% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_genotypes.R
\name{recode_genotypes}
\alias{recode_genotypes}
\title{Recode Numeric Genotype Data to Character Labels}
\usage{
recode_genotypes(genotype_data, id_column = "probeset_id")
}
\arguments{
\item{genotype_data}{A data frame containing genotype data with a probeset_id column
and numeric genotype columns coded as -1, 0, 1, or 2}

\item{id_column}{Character string specifying the name of the ID column to exclude
from recoding (default: "probeset_id")}
}
\value{
A data frame with the same structure as input, but with numeric genotype
values recoded to character labels
}
\description{
This function converts numeric genotype codes to their corresponding character labels:
-1 -> "NoCall", 0 -> "AA", 1 -> "AB", 2 -> "BB"
}
\examples{
# Create sample data
sample_data <- data.frame(
  probeset_id = c("SNP1", "SNP2", "SNP3"),
  sample1 = c(-1, 0, 1),
  sample2 = c(0, 1, 2),
  sample3 = c(2, -1, 0)
)

# Recode genotypes
recoded_data <- recode_genotypes(sample_data)
print(recoded_data)

}
