% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPutils-package.R
\docType{package}
\name{SNPutils-package}
\alias{SNPutils}
\alias{SNPutils-package}
\title{SNPutils: Convert Genotype Data to Allele Formats and Generate Summary Statistics}
\description{
Tools for converting numerical genotype data (0, 1, 2) to allele pairs, generating TYP and COM file formats, validating allele assignments, and calculating sample and variant summary statistics.
}
\author{
\strong{Maintainer}: Hedia Tnani \email{hediatnani0@gmail.com} (\href{https://orcid.org/0000-0002-0380-9740}{ORCID})

}
\keyword{internal}
